Refactor status enums and improve score precision
- Introduced AccountStatus and QuizAttemptStatus enums
- Updated User and QuizAttempt to use enums for Status
- Changed PassScore and Score to decimal for accuracy
- Made StudentDiplomaEnrollment inherit from Entity
- Cleaned up csproj file and removed merge markers
- Removed unused constructor from GenericRepository11

+++ ExaminationSystem/ExaminationSystem.Domin/Entities/Quiz.cs
@@ -20,7 +20,7 @@ namespace ExaminationSystem.Entities
 
         [Required]
         [Range(0, 100)]
-        public int PassScore { get; set; } = 60;
+        public decimal PassScore { get; set; } = 60;
 
         public int? MaxAttempts { get; set; }
